Updating mana and health values server side

+++ core/src/ee/taltech/game/server/messages/UpdateMana.java
@@ -3,7 +3,7 @@ package ee.taltech.game.server.messages;
 public class UpdateMana {
 
     public Integer playerId;
-    public Integer mana;
+    public double mana;
 
     /**
      * Empty constructor for Kryonet.
@@ -16,7 +16,7 @@ public class UpdateMana {
      * @param id player's ID, who's health is updated
      * @param mana new mana of the player
      */
-    public UpdateMana(Integer id, Integer mana) {
+    public UpdateMana(Integer id, double mana) {
         this.playerId = id;
         this.mana = mana;
     }
